+++ src/HasteModuleLoader/HasteModuleLoader.js
@@ -21,49 +21,26 @@ const transform = require('../lib/transform');
 const utils = require('../lib/utils');
 
 const COVERAGE_STORAGE_VAR_NAME = '____JEST_COVERAGE_DATA____';
-const NODE_PATH = process.env.NODE_PATH;
+const NODE_PATH =
+  (process.env.NODE_PATH ? process.env.NODE_PATH.split(path.delimiter) : null);
 const IS_PATH_BASED_MODULE_NAME = /^(?:\.\.?\/|\/)/;
 const VENDOR_PATH = path.resolve(__dirname, '../../vendor');
-const NODE_CORE_MODULES = {
-  assert: true,
-  buffer: true,
-  child_process: true,
-  cluster: true,
-  console: true,
-  constants: true,
-  crypto: true,
-  dgram: true,
-  dns: true,
-  domain: true,
-  events: true,
-  freelist: true,
-  fs: true,
-  http: true,
-  https: true,
-  module: true,
-  net: true,
-  os: true,
-  path: true,
-  punycode: true,
-  querystring: true,
-  readline: true,
-  repl: true,
-  smalloc: true,
-  stream: true,
-  string_decoder: true,
-  sys: true,
-  timers: true,
-  tls: true,
-  tty: true,
-  url: true,
-  util: true,
-  vm: true,
-  zlib: true,
-};
 
 const mockParentModule = {
   id: 'mockParent',
   exports: {},
+};
+
+const isFile = file => {
+  let stat;
+  try {
+    stat = fs.statSync(file);
+  } catch (err) {
+    if (err && err.code === 'ENOENT') {
+      return false;
+    }
+  }
+  return stat.isFile() || stat.isFIFO();
 };
 
 const hasOwnProperty = Object.prototype.hasOwnProperty;
@@ -119,6 +96,7 @@ class Loader {
     this._reverseDependencyMap = null;
     this._shouldAutoMock = true;
     this._configShouldMockModuleNames = {};
+    this._extensions = config.moduleFileExtensions.map(ext => '.' + ext);
 
     if (config.collectCoverage) {
       this._CoverageCollector = require(config.coverageCollector);
@@ -335,7 +313,7 @@ class Loader {
     let mockAbsPath = null;
     let realAbsPath = null;
 
-    if (hasOwnProperty.call(NODE_CORE_MODULES, moduleName)) {
+    if (resolve.isCore(moduleName)) {
       moduleType = 'node';
       realAbsPath = moduleName;
     } else {
@@ -350,7 +328,7 @@ class Loader {
         const absolutePath = this._moduleNameToPath(currPath, moduleName);
         if (absolutePath === undefined) {
           throw new Error(
-            `Cannot find module '${moduleName}' from '${currPath}'`
+            `Cannot find module '${moduleName}' from '${currPath || '.'}'`
           );
         }
 
@@ -390,74 +368,20 @@ class Loader {
     return moduleID.split(path.delimiter)[1];
   }
 
-  /**
-   * Given a module name and the current file path, returns the normalized
-   * (absolute) module path for said module. Relative-path CommonJS require()s
-   * such as `require('./otherModule')` need to be looked up with context of
-   * the module that's calling require()
-   *
-   * Also contains special case logic for built-in modules, in which it just
-   * returns the module name.
-   */
+
   _moduleNameToPath(currPath, moduleName) {
-    // Relative-path CommonJS require()s such as `require('./otherModule')`
-    // need to be looked up with context of the module that's calling
-    // require().
-    if (IS_PATH_BASED_MODULE_NAME.test(moduleName)) {
-      // Normalize the relative path to an absolute path
-      const modulePath = path.resolve(currPath, '..', moduleName);
-      const extensions = this._config.moduleFileExtensions;
-
-      // http://nodejs.org/docs/v0.10.0/api/all.html#all_all_together
-      // LOAD_AS_FILE #1
-      if (fs.existsSync(modulePath) && fs.statSync(modulePath).isFile()) {
-        return modulePath;
-      }
-      // LOAD_AS_FILE #2+
-      for (let i = 0; i < extensions.length; i++) {
-        const fullPath = modulePath + '.' + extensions[i];
-        if (fs.existsSync(fullPath) && fs.statSync(fullPath).isFile()) {
-          return fullPath;
-        }
-      }
-      // LOAD_AS_DIRECTORY
-      if (fs.existsSync(modulePath) && fs.statSync(modulePath).isDirectory()) {
-        // LOAD_AS_DIRECTORY #1
-        const packagePath = path.join(modulePath, 'package.json');
-        if (fs.existsSync(packagePath)) {
-          const packageData = require(packagePath);
-          if (packageData.main) {
-            const mainPath = path.join(modulePath, packageData.main);
-            if (fs.existsSync(mainPath)) {
-              return mainPath;
-            }
-          }
-        }
-
-        // The required path is a valid directory, but there's no matching
-        // js file at the same path. So look in the directory for an
-        // index.js file.
-        const indexPath = path.join(modulePath, 'index');
-        for (let i = 0; i < extensions.length; i++) {
-          const fullPath = indexPath + '.' + extensions[i];
-          if (fs.existsSync(fullPath) && fs.statSync(fullPath).isFile()) {
-            return fullPath;
-          }
-        }
-      }
-    } else {
-      const resource = this._getResource('JS', moduleName);
-      if (!resource) {
-        return this._nodeModuleNameToPath(currPath, moduleName);
-      }
-      return resource.path;
+    const resource = this._getResource('JS', moduleName);
+    if (!resource) {
+      return this._nodeModuleNameToPath(currPath, moduleName);
     }
+    return resource.path;
   }
 
   _nodeModuleNameToPath(currPath, moduleName) {
     // Handle module names like require('jest/lib/util')
     let subModulePath = null;
     let moduleProjectPart = moduleName;
+    const basedir = path.dirname(currPath);
     if (/\//.test(moduleName)) {
       const projectPathParts = moduleName.split('/');
       moduleProjectPart = projectPathParts.shift();
@@ -465,28 +389,20 @@ class Loader {
     }
 
     let resolveError = null;
-    const exts = this._config.moduleFileExtensions.map(ext => '.' + ext);
     try {
-      if (NODE_PATH) {
-        return resolve.sync(moduleName, {
-          paths: NODE_PATH.split(path.delimiter),
-          basedir: path.dirname(currPath),
-          extensions: exts,
-        });
-      } else {
-        return resolve.sync(moduleName, {
-          basedir: path.dirname(currPath),
-          extensions: exts,
-        });
-      }
+      return resolve.sync(moduleName, {
+        basedir,
+        extensions: this._extensions,
+        isFile,
+        paths: NODE_PATH,
+        readFileSync: fs.readFileSync,
+      });
     } catch (e) {
-      // Facebook has clowny package.json resolution rules that don't apply to
-      // regular Node rules. Until we can make ModuleLoaders more pluggable
-      // (so that FB can have a custom ModuleLoader and all the normal people
-      // can have a normal ModuleLoader), we catch node-resolution exceptions
-      // and fall back to some custom resolution logic before throwing the
-      // error.
-      resolveError = e;
+      // resolve.sync uses the basedir instead of currPath and therefore
+      // doesn't throw an accurate error message.
+      resolveError = new Error(
+        `Cannot find module '${moduleName}' from '${currPath || '.'}'`
+      );
     }
 
     // Memoize the project name -> package.json resource lookup map
@@ -509,8 +425,7 @@ class Loader {
     // Make sure the resource path is above the currPath in the fs path
     // tree. If so, just use node's resolve
     const resourceDirname = path.dirname(resource.path);
-    const currFileDirname = path.dirname(currPath);
-    if (resourceDirname.indexOf(currFileDirname) > 0) {
+    if (resourceDirname.indexOf(basedir) > 0) {
       throw resolveError;
     }
 
@@ -533,7 +448,7 @@ class Loader {
     const moduleID = this._getNormalizedModuleID(currPath, moduleName);
     if (hasOwnProperty.call(this._explicitShouldMock, moduleID)) {
       return this._explicitShouldMock[moduleID];
-    } else if (NODE_CORE_MODULES[moduleName]) {
+    } else if (resolve.isCore(moduleName)) {
       return false;
     } else if (this._shouldAutoMock) {
       // See if the module is specified in the config as a module that should
@@ -592,21 +507,19 @@ class Loader {
   }
 
   constructBoundRequire(modulePath) {
-    const boundModuleRequire = this.requireModuleOrMock.bind(this, modulePath);
+    const moduleRequire = this.requireModuleOrMock.bind(this, modulePath);
 
-    boundModuleRequire.resolve = moduleName => {
+    moduleRequire.resolve = moduleName => {
       const ret = this._moduleNameToPath(modulePath, moduleName);
       if (!ret) {
         throw new Error(`Module(${moduleName}) not found!`);
       }
       return ret;
     };
-    boundModuleRequire.generateMock = this._generateMock.bind(this, modulePath);
-    boundModuleRequire.requireMock = this.requireMock.bind(this, modulePath);
-    boundModuleRequire.requireActual =
-      this.requireModule.bind(this, modulePath);
+    moduleRequire.requireMock = this.requireMock.bind(this, modulePath);
+    moduleRequire.requireActual = this.requireModule.bind(this, modulePath);
 
-    return boundModuleRequire;
+    return moduleRequire;
   }
 
   /**
@@ -654,8 +567,10 @@ class Loader {
       throw new Error(`Unknown modulePath: ${modulePath}`);
     }
 
-    if (resource.type === 'ProjectConfiguration'
-        || resource.type === 'Resource') {
+    if (
+      resource.type === 'ProjectConfiguration' ||
+      resource.type === 'Resource'
+    ) {
       throw new Error(
         `Could not extract dependency information from this type of file!`
       );
@@ -801,7 +716,7 @@ class Loader {
       modulePath = manualMockResource.path;
     }
 
-    if (NODE_CORE_MODULES[moduleName]) {
+    if (resolve.isCore(moduleName)) {
       return require(moduleName);
     }